api/requirements.txt
fastapi
strawberry-graphql[fastapi]
